pyproject.toml
[project]
name = "Passman - Password Manager"
version = "0.1"
description = "Didactic Password Manager"
authors = [
    {name = "Cédric Vanconingsloo", email = "cedric.vanconingsloo@ifosup.wavre.be"},
]
dependencies = [
    "black>=22.8.0",
    "pillow>=9.2.0",
]
requires-python = ">=3.10,<3.11"
license = {text = "GPL"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.1.3",
]
